+++ src/i_timer.h
@@ -40,10 +40,10 @@
 
 // Called by D_DoomLoop,
 // returns current time in tics.
-int I_GetTime(void);
+uint64_t I_GetTime(void);
 
 // returns current time in ms
-int I_GetTimeMS(void);
+uint64_t I_GetTimeMS(void);
 
 // Pause for a specified number of ms
 void I_Sleep(int ms);
